Remove deprecated code related to events
