+++ src/apps/companies/apps/activity-feed/constants.js
@@ -9,58 +9,6 @@ const CONTACT_ACTIVITY_SORT_SELECT_OPTIONS = [
   { name: 'Oldest', value: 'date:asc' },
 ]
 
-const EVENT_ATTENDEES_SORT_OPTIONS = {
-  'first_name:asc': {
-    'object.dit:firstName': {
-      order: 'asc',
-      unmapped_type: 'string',
-    },
-  },
-  'first_name:desc': {
-    'object.dit:firstName': {
-      order: 'desc',
-      unmapped_type: 'string',
-    },
-  },
-  'last_name:asc': {
-    'object.dit:lastName': {
-      order: 'asc',
-      unmapped_type: 'string',
-    },
-  },
-  'last_name:desc': {
-    'object.dit:lastName': {
-      order: 'desc',
-      unmapped_type: 'string',
-    },
-  },
-  'company_name:asc': {
-    'object.dit:companyName': {
-      order: 'asc',
-      unmapped_type: 'string',
-    },
-  },
-  'company_name:desc': {
-    'object.dit:companyName': {
-      order: 'desc',
-      unmapped_type: 'string',
-    },
-  },
-}
-
-const EVENT_AVENTRI_ATTENDEES_STATUS = {
-  activated: 'Activated',
-  attended: 'Attended',
-  confirmed: 'Confirmed',
-  cancelled: 'Cancelled',
-  noShow: 'No Show',
-  waitlist: 'Waitlist',
-}
-
-const EVENT_AVENTRI_ATTENDEES_STATUSES = Object.values(
-  EVENT_AVENTRI_ATTENDEES_STATUS
-)
-
 const EVENT_ATTENDEES_STATUS = {
   registered: 'Registered',
   waitingList: 'Waiting list',
@@ -68,32 +16,6 @@ const EVENT_ATTENDEES_STATUS = {
   attended: 'Attended',
   cancelled: 'Cancelled',
 }
-
-const createMappingObject = (status, aventriStatuses) => ({
-  [status]: {
-    statuses: aventriStatuses,
-    urlSlug: status.replaceAll(' ', '-').toLowerCase(),
-  },
-})
-
-const EVENT_ATTENDEES_MAPPING = [
-  createMappingObject(EVENT_ATTENDEES_STATUS.registered, [
-    EVENT_AVENTRI_ATTENDEES_STATUS.activated,
-    EVENT_AVENTRI_ATTENDEES_STATUS.confirmed,
-  ]),
-  createMappingObject(EVENT_ATTENDEES_STATUS.cancelled, [
-    EVENT_AVENTRI_ATTENDEES_STATUS.cancelled,
-  ]),
-  createMappingObject(EVENT_ATTENDEES_STATUS.attended, [
-    EVENT_AVENTRI_ATTENDEES_STATUS.attended,
-  ]),
-  createMappingObject(EVENT_ATTENDEES_STATUS.didNotAttend, [
-    EVENT_AVENTRI_ATTENDEES_STATUS.noShow,
-  ]),
-  createMappingObject(EVENT_ATTENDEES_STATUS.waitingList, [
-    EVENT_AVENTRI_ATTENDEES_STATUS.waitlist,
-  ]),
-].reduce((a, v) => ({ ...a, ...v }), {})
 
 const DATA_HUB_ACTIVITY = [
   'dit:Interaction', // Interaction
@@ -111,13 +33,9 @@ const EXTERNAL_ACTIVITY = [
 ]
 
 module.exports = {
-  EVENT_ATTENDEES_SORT_OPTIONS,
-  EVENT_AVENTRI_ATTENDEES_STATUS,
-  EVENT_AVENTRI_ATTENDEES_STATUSES,
   FILTER_FEED_TYPE,
   DATA_HUB_ACTIVITY,
   EXTERNAL_ACTIVITY,
   CONTACT_ACTIVITY_SORT_SELECT_OPTIONS,
   EVENT_ATTENDEES_STATUS,
-  EVENT_ATTENDEES_MAPPING,
 }
