--- src/apps/companies/apps/activity-feed/es-queries/aventri-event-query.js
@@ -1,22 +1,0 @@
-const aventriEventQuery = (eventIds) => ({
-  query: {
-    bool: {
-      must: [
-        {
-          term: {
-            'object.type': 'dit:aventri:Event',
-          },
-        },
-        {
-          terms: {
-            id: eventIds,
-          },
-        },
-      ],
-    },
-  },
-})
-
-module.exports = {
-  aventriEventQuery,
-}
