+++ src/apps/companies/apps/activity-feed/repos.js
@@ -11,14 +11,6 @@ async function fetchActivityFeed(req, body) {
 }
 
 // istanbul ignore next: Covered by functional tests
-function fetchMatchingDataHubContact(req, attendeeEmail) {
-  // istanbul ignore next: Covered by functional tests
-  return authorisedRequest(req, {
-    method: 'GET',
-    url: `${config.apiRoot}/v3/contact?email=${attendeeEmail}`,
-  })
-}
-// istanbul ignore next: Covered by functional tests
 async function fetchUserFeatureFlags(req) {
   // istanbul ignore next: Covered by functional tests
   const user = await authorisedRequest(req, `${config.apiRoot}/whoami/`)
@@ -26,6 +18,5 @@ async function fetchUserFeatureFlags(req) {
 }
 module.exports = {
   fetchActivityFeed,
-  fetchMatchingDataHubContact,
   fetchUserFeatureFlags,
 }
