+++ src/apps/events/router.js
@@ -1,7 +1,5 @@
 const router = require('express').Router()
 
-const urls = require('../../lib/urls')
-
 const { APP_PERMISSIONS, LOCAL_NAV } = require('./constants')
 
 const { handleRoutePermissions, setLocalNav } = require('../middleware')
@@ -9,21 +7,11 @@ const { getEventDetails } = require('./middleware/details')
 const { renderEventsView } = require('./controllers/events')
 const attendeesRouter = require('./attendees/router')
 
-const {
-  fetchAventriEvent,
-  fetchAventriEventRegistrationStatusAttendees,
-} = require('../companies/apps/activity-feed/controllers')
-
 router.get('/create', renderEventsView)
 router.use(handleRoutePermissions(APP_PERMISSIONS))
 
 router.get('/aventri/:aventriEventId/details', renderEventsView)
-router.get(urls.events.aventri.detailsData.route, fetchAventriEvent)
 router.get('/aventri/:aventriEventId/registration/:status', renderEventsView)
-router.get(
-  urls.events.aventri.registrationStatusData.route,
-  fetchAventriEventRegistrationStatusAttendees
-)
 
 router.use(
   '/:eventId',
