--- test/unit/data/activity-feed/aventri-registration-status-no-details.json
@@ -1,25 +1,0 @@
-{
-  "took" : 1009,
-  "timed_out" : false,
-  "_shards" : {
-    "total" : 378,
-    "successful" : 378,
-    "skipped" : 220,
-    "failed" : 0
-  },
-  "hits" : {
-    "total" : {
-      "value" : 0,
-      "relation" : "eq"
-    },
-    "max_score" : null,
-    "hits" : [ ]
-  },
-  "aggregations" : {
-    "countfield" : {
-      "doc_count_error_upper_bound" : 0,
-      "sum_other_doc_count" : 0,
-      "buckets" : [ ]
-    }
-  }
-}
