--- test/unit/data/activity-feed/aventri-registration-status-with-aggregation-counts.json
@@ -1,54 +1,0 @@
-{
-  "took" : 965,
-  "timed_out" : false,
-  "_shards" : {
-    "total" : 274,
-    "successful" : 274,
-    "skipped" : 142,
-    "failed" : 0
-  },
-  "hits" : {
-    "total" : {
-      "value" : 38,
-      "relation" : "eq"
-    },
-    "max_score" : null,
-    "hits" : [ ]
-  },
-  "aggregations" : {
-    "countfield" : {
-      "doc_count_error_upper_bound" : 0,
-      "sum_other_doc_count" : 0,
-      "buckets" : [
-        {
-          "key" : "Attended",
-          "doc_count" : 24
-        },
-        {
-          "key" : "Activated",
-          "doc_count" : 4
-        },
-        {
-          "key" : "Waitlist",
-          "doc_count" : 4
-        },
-        {
-          "key" : "Confirmed",
-          "doc_count" : 0
-        },
-        {
-          "key" : "Incomplete",
-          "doc_count" : 2
-        },
-        {
-          "key" : "Cancelled",
-          "doc_count" : 7
-        },
-        {
-          "key" : "No Show",
-          "doc_count" : 2
-        }
-      ]
-    }
-  }
-}
